fixes narrowing type conversion

+++ cub/cub/device/dispatch/dispatch_merge_sort.cuh
@@ -626,7 +626,7 @@ struct DispatchMergeSort : SelectedPolicy
 
       // Invoke DeviceMergeSortBlockSortKernel
       THRUST_NS_QUALIFIER::cuda_cub::launcher::triple_chevron(
-        num_tiles, merge_sort_helper_t::policy_t::BLOCK_THREADS, 0, stream)
+        static_cast<int>(num_tiles), merge_sort_helper_t::policy_t::BLOCK_THREADS, 0, stream)
         .doit(
           DeviceMergeSortBlockSortKernel<
             MaxPolicyT,
